requirements-py3.txt
atomicwrites==1.2.1
attrs==18.2.0
autopep8==1.2.4
backports.functools-lru-cache==1.5
beautifulsoup4==4.7.0
certifi==2018.11.29
distlib==0.2.8
django-debug-toolbar==1.4
django-dynamic-fixture==1.9.0
flake8==2.5.4
idna==2.8
isort==4.2.5
lxml==4.2.5
mccabe==0.4.0
MechanicalSoup==0.11.0
mock==2.0.0
more-itertools==5.0.0
mysqlclient==1.4.6
packaging==18.0
pathlib2==2.3.3
pbr==5.1.1
pep8==1.7.1
pluggy==0.8.0
progressbar==2.5
py==1.7.0
py3dns==3.2.0
pyflakes==1.0.0
pyparsing==2.3.0
pytest==4.0.2
pytest-cov==2.6.0
pytest-django==2.9.1
pytest-pythonpath==0.7
requests==2.21.0
soupsieve==1.6
urllib3==1.24.1
django-cors-headers==3.2.0
django-ckeditor==5.8.0
python-memcached==1.59
celery>=4.4.0,<4.5.0
pysrt>=1.1.2,<1.2.0
django-simple-history>=2.10.0,<2.11.0
